+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -487,7 +487,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYPCGWYT9R33C68038X8RAW8"),
+  _id: z.string().default("01KYPP1H6E3DSJGNG5T58YX4J5"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: CreateDatasourceStatus$inboundSchema,