+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -187,7 +187,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYPCGWYSM9CR7AYSRMM78K1G"),
+  _id: z.string().default("01KYPP1H6DSR6D22QX6RZXQKEP"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: RetrieveDatasourceStatus$inboundSchema,