+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -221,7 +221,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYPCGWYXNF8PEETM0DY5C98Q"),
+  _id: z.string().default("01KYPP1H6HN2QT9QB97H7TK9PX"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: UpdateDatasourceStatus$inboundSchema,