+++ packages/orq-rc/RELEASES.md
@@ -11229,3 +11229,13 @@ Based on:
 - [typescript v4.13.0-rc.13] packages/orq-rc
 ### Releases
 - [NPM v4.13.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.13 - packages/orq-rc
+
+## 2026-07-29 10:20:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.13.0-rc.14] packages/orq-rc
+### Releases
+- [NPM v4.13.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.14 - packages/orq-rc